--- types/ca_paths.go
@@ -1,9 +1,0 @@
-package types
-
-type CaPaths interface {
-	NodeCertKeyAbsFilename(identifier string) string
-	NodeCertAbsFilename(identifier string) string
-	NodeCertCSRAbsFilename(identifier string) string
-	CANodeDir(string) string
-	RootCaIdentifier() string
-}
